Cleaned columns
I removed a lot of lines of my code because they either did not work or was not needed. I then spend some time researching how to create a .csv file from a df which I used to create the cleaned df. The code worked as intended as I now have a .csv file without the useless columns. Now I need to remove all the cities other than Sydney.